Better time management on Linux boxes
Uses timing with nanosecond precision.

+++ ansmatrix.h
@@ -1,4 +1,4 @@
 #ifndef ANSMATRIX_H
 #define ANSMATRIX_H
-void ansmatrix(unsigned short B[20][20]);
+void ansmatrix(unsigned long int B[20][20]);
 #endif
